+++ R/createDT.R
@@ -2,19 +2,22 @@
 #'
 #' Generate an interactive data table with download buttons.
 #'
-#' @param DF Data to show.
+#' @param dat Data to show.
 #' @param caption Table caption.
 #' @param scrollY Maximum height (in pixels) before defaulting to scrolling.
 #' 
 #' @family general
 #' @export
 #' @importFrom DT datatable
-createDT <- function(DF, 
+#' @examples  
+#' echodata::createDT(dat = echodata::BST1)
+createDT <- function(dat, 
                      caption="", 
                      scrollY=400){
     data <- DT::datatable(
-        DF, caption=caption,
-        extensions = 'Buttons',
+      data = dat, 
+      caption=caption,
+      extensions = 'Buttons',
       options = list( dom = 'Bfrtip',
                       buttons = c('copy', 'csv', 'excel', 'pdf', 'print'),
                       scrollY = scrollY,